requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==2.1.1
Deprecated==1.2.18
Django==4.2
django-braces==1.15.0
django-cors-headers==4.6.0
django-debug-toolbar==3.6.0
django-embed-video==1.4.4
django-redisboard==8.3.0
djangorestframework==3.15.2
idna==3.10
packaging==24.2
Pillow==9.2.0
pymemcache==3.5.2
pytz==2025.1
redis==4.3.4
requests==2.28.1
six==1.17.0
sqlparse==0.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
wrapt==1.17.2
